chore: remove fixed versions for dependencies (#34)
* feat(smt): add rocksdb support (#9)

This commit introduces RocksDB as backend for the SMT.

closes #5

* fix(smt): allow value replace (#13)

* feat(smt): add rocksdb support (#9) (#10)

* fix(smt): allow value replace

This commit fixes the SMT behavior that will iterate over the bits of
the key until depletion. It instead replaces the value upon depletion.

It assumes the hash cannot produce more bits than [HASH_LEN] so a
collision is necessarily the same data being inserted.

related to #12

* improve docs

* feat: add execution context

* remove clippy check for main

* change deref due to a compiler bug

* chore: remove fixed versions for dependencies

* update names to glossary & version to 0.1.4